Added Code To Work With Windows

+++ library/wpos/models/WposSocketControl.php
@@ -29,8 +29,12 @@ class WposSocketControl {
      * @return mixed API result array
      */
     public function startSocketServer($result){
-        exec("nodejs ".$_SERVER['DOCUMENT_ROOT'].$_SERVER['APP_ROOT']."api/server.js > /dev/null & echo $!", $output);
-        if ($this->getServerStat()===false){
+		if (strtoupper(substr(PHP_OS, 0, 3)) === 'WIN') {
+			pclose(popen('START "WPOS" node '.$_SERVER['DOCUMENT_ROOT'].$_SERVER['APP_ROOT'].'api/server.js','r'));
+		} else {
+			exec("nodejs ".$_SERVER['DOCUMENT_ROOT'].$_SERVER['APP_ROOT']."api/server.js > /dev/null & echo $!", $output);
+        }
+		if ($this->getServerStat()===false){
             $result['error'] = "Failed to start the feed server!";
         }
         return $result;
@@ -42,7 +46,11 @@ class WposSocketControl {
      * @return mixed API result array
      */
     public function stopSocketServer($result){
-        exec("kill `pidof nodejs`", $output);
+		if (strtoupper(substr(PHP_OS, 0, 3)) === 'WIN') {
+			exec('TASKKILL /F /FI "WindowTitle eq WPOS"', $output);
+		} else {
+			exec("kill `pidof nodejs`", $output);
+		}
         if ($this->getServerStat()===true){
             $result['error'] = "Failed to stop the feed server!";
         }
@@ -85,17 +93,28 @@ class WposSocketControl {
      * @return bool
      */
     private function getServerStat(){
-        exec('ps cax | grep nodejs > /dev/null
-                if [ $? -eq 0 ]; then
-                    echo "Online"
-                        else
-                    echo "Offline"
-                fi', $output);
-        if (strpos($output[0], 'Online')!==false){
-            return true;
-        } else {
-            return false;
-        }
+		if (strtoupper(substr(PHP_OS, 0, 3)) === 'WIN') {
+			exec('TASKLIST /NH /V /FI "WindowTitle eq WPOS"', $output );
+			if (strpos($output[0], 'INFO')!==false){
+				$output[0] = 'Offline';
+				return false;
+			} else {
+				$output[0] = 'Online';
+				return true;
+			}
+		} else {
+			exec('ps cax | grep nodejs > /dev/null
+					if [ $? -eq 0 ]; then
+						echo "Online"
+							else
+						echo "Offline"
+					fi', $output);
+			if (strpos($output[0], 'Online')!==false){
+				return true;
+			} else {
+				return false;
+			}
+		}
     }
 
 }
